app/requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
orjson==3.6.7
gunicorn==20.1.0
requests==2.27.1
python-multipart==0.0.5
mysql-connector-python==8.0.28
